Switch reactions position on PostCard
