+++ src/components/EmojiPill/EmojiPill.scss
@@ -3,7 +3,7 @@
   text-transform: none;
   font-weight: 400;
   line-height: 18px;
-  margin-bottom: 0px;
+  margin-bottom: 2px;
   font-size: 1em;
   cursor: pointer;
 }
