+++ src/components/EmojiRow/EmojiRow.scss
@@ -1,0 +1,9 @@
+.footer-reactions{
+  transition: all .25s ease;
+  display: flex;
+  position: relative;
+  align-items: center;
+  margin-bottom: 8px;
+  height: auto;
+  flex-wrap: wrap;
+}
