+++ src/components/EmojiRow/index.js
@@ -1,0 +1,2 @@
+import component from './EmojiRow'
+export default component
